Current endpoints working, changed urls to be more RESTful
